Moved _get_node_id() to orchestrator class to avoid get it called multiple times
